Replace old category menu with the new one
