+++ src/containers/CategoryMenu/CategoryMenu.tsx
@@ -1,85 +1,72 @@
-import React, { useEffect } from 'react';
-import { bindActionCreators, Dispatch } from 'redux';
-import { RouterState, routerActions as cRouterActions } from 'connected-react-router';
-import { connect } from 'react-redux';
+import React from 'react';
 import { Row, Col } from 'react-flexbox-grid';
 import classNames from 'classnames';
-import _ from 'lodash';
-import queryString from 'query-string';
-import { productActions as cProductActions } from '../../actions/products';
-import { categoryActions as cCategoryActions } from '../../actions/categories';
-import { userActions as cUserActions } from '../../actions/users';
+import useSWR from 'swr';
+import { useHistory, useLocation } from 'react-router-dom';
+import { useQueryParams, StringParam } from 'use-query-params';
 import { isAdminPage } from '../../modules/helpers';
 import styles from './CategoryMenu.module.scss';
+import { fetchCategories } from '../../api';
+import { clearSession } from '../../modules/session';
 
-interface StateProps {
-  categories: CategoriesState;
-  router: RouterState;
-}
+const CategoryMenu = () => {
+  const [query, setQuery] = useQueryParams({ categoryId: StringParam });
 
-interface DispatchProps {
-  categoryActions: typeof cCategoryActions;
-  productActions: typeof cProductActions;
-  routerActions: typeof cRouterActions;
-  userActions: typeof cUserActions;
-}
+  const history = useHistory();
+  const location = useLocation();
 
-type CategoryMenuProps = StateProps & DispatchProps;
+  const { data: categories } = useSWR('/categories', fetchCategories);
 
-const CategoryMenu = ({
-  categories,
-  router,
-  categoryActions,
-  productActions,
-  routerActions,
-  userActions,
-}: CategoryMenuProps) => {
-  useEffect(() => {
-    categoryActions.fetchCategories();
-  }, [categoryActions]);
+  const isRoot = location.pathname === '/';
 
-  useEffect(() => {
-    if (router.location.search) {
-      productActions.fetchProducts();
+  const navigateToCategory = (categoryId?: string) => {
+    if (!isRoot) {
+      history.push('/');
     }
-  }, [router.location.search, productActions]);
+    setQuery({ categoryId });
+  };
 
-  const query = queryString.parse(router.location.search);
-  const isRoot = router.location.pathname === '/';
   return (
     <Row center="xs" className={styles.menu}>
       <Col xs={12} lg={8}>
         <div className={styles.itemsWrapper}>
-          {!isAdminPage(router.location.pathname) && (
+          {!isAdminPage(window.location.pathname) && (
             <>
               <div className={styles.categories}>
                 <div
                   className={classNames(styles.menuItem, {
-                    [styles.activeItem]: isRoot && !query.category,
+                    [styles.activeItem]: isRoot && !query.categoryId,
                   })}
-                  onClick={() => routerActions.push(`/`)}
+                  onClick={() => navigateToCategory(undefined)}
                 >
                   Home
                 </div>
-                {_.map(categories.data, (category, index) => (
-                  <div
-                    key={index}
-                    className={classNames(styles.menuItem, {
-                      [styles.activeItem]: isRoot && query.category === category._id,
-                    })}
-                    onClick={() => routerActions.push(`/?category=${category._id}`)}
-                  >
-                    {category.name}
-                  </div>
-                ))}
+                {categories &&
+                  categories.map(category => (
+                    <div
+                      key={category._id}
+                      className={classNames(styles.menuItem, {
+                        [styles.activeItem]: isRoot && query.categoryId === category._id,
+                      })}
+                      onClick={() => navigateToCategory(category._id)}
+                    >
+                      {category.name}
+                    </div>
+                  ))}
               </div>
-              <div className={styles.menuItem} onClick={() => routerActions.push('/about')}>
+              <div className={styles.menuItem} onClick={() => history.push('/about')}>
                 Contato
               </div>
             </>
           )}
-          {isAdminPage(router.location.pathname) && (
-            <div className={styles.menuItem} onClick={userActions.logout}>
+          {isAdminPage(window.location.pathname) && (
+            <div
+              className={styles.menuItem}
+              onClick={() => {
+                clearSession();
+                history.push('/');
+              }}
+            >
               Logout
             </div>
           )}
@@ -89,16 +76,4 @@ const CategoryMenu = ({
   );
 };
 
-const mapStateToProps = (state: RootState) => ({
-  categories: state.categories,
-  router: state.router,
-});
-
-const mapDispatchToProps = (dispatch: Dispatch) => ({
-  categoryActions: bindActionCreators({ ...cCategoryActions }, dispatch),
-  productActions: bindActionCreators({ ...cProductActions }, dispatch),
-  routerActions: bindActionCreators({ ...cRouterActions }, dispatch),
-  userActions: bindActionCreators({ ...cUserActions }, dispatch),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(CategoryMenu);
+export default CategoryMenu;
