--- src/containers/NewCategoryMenu/NewCategoryMenu.module.scss
@@ -1,37 +1,0 @@
-@import '../../assets/colors';
-
-.menu {
-  background-color: $cherry-blossom;
-  color: $body-font-color;
-  height: 32px;
-  box-shadow: $box-shadow;
-  border-bottom: 1px solid $light-silver;
-
-  .itemsWrapper {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 32px;
-    height: 100%;
-
-    .categories {
-      display: flex;
-    }
-
-    .menuItem {
-      cursor: pointer;
-      color: $body-font-color;
-      margin-right: 16px;
-
-      &:hover {
-        color: $cadillac;
-      }
-
-      &.activeItem {
-        font-weight: 600;
-        color: $cadillac;
-        text-decoration: underline;
-      }
-    }
-  }
-}
